requirements.txt
flask==3.0.0
flask-cors==5.0.0
tensorflow==2.16.1
numpy==1.26.1
pytesseract==0.3.10
pillow==10.2.0
easyocr==1.7.2
